Add a launcher shared library
To share bits of the launcher that the firmware also uses. One of the
headers is _also_ used by the HAL...
