+++ CMakeLists.txt
@@ -9,6 +9,8 @@ endif()
 include(32blit.cmake)
 add_subdirectory(examples)
 
+add_subdirectory(launcher-shared)
+
 if(${CMAKE_SYSTEM_NAME} STREQUAL Generic)
     add_subdirectory(32blit-stm32)
     add_subdirectory(firmware)
