+++ launcher-shared/executable.hpp
